src/requirements.txt
pynput~=1.7.3
pywin32
attrs~=20.3.0
opencv-python~=4.5.1.48
numpy~=1.20.2
Pillow~=8.2.0
pyyaml
